demo3/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.3.0
tqdm>=4.62.0
pyyaml>=5.4.0
einops>=0.3.0
pytorch-msssim>=0.2.0
opencv-python>=4.5.0
scikit-image>=0.18.0
matplotlib>=3.3.0
pandas>=1.3.0
seaborn>=0.11.0
tensorboard>=2.7.0